Adds solution for day 13 part 12
Part one solution had a bug, that didn't allow it to be reused. Instead
opting for a proper implementation of breadth first, a'la
https://en.wikipedia.org/wiki/Breadth-first_search, which in turn
requires a queue datastructure. Included with this solution is a linked
list implementation, which can readily act as queue, since it has
operations to add/remove elements from both head and tail.