Fixes a subtle bug in 17p2
Instead of only moving to the exit, when that is the only possible move,
detect the case where a move to the exit is possible amongst others, but
all other moves leads to dead ends. Hence choosing to move to the exit
at that point is the right choice.

---

Also a bit of cleanup and renaming of sp to lp.