Updates day 5 part 2
This commit extracts the intcode computer into a standalone unit,
callable via `intcode_run()`.

A further change is introduced into the main program, which now takes
care of providing the input the the intcode computer via stdin, instead
of having it be provided from from echo via the Makefile target.

To do this a fork/pipe approach is introduced, where a child process is
forked off to run the actual intcode computer program, and then the
parent process sends input via a pipe that is attached to stdin of the
child.